% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Banco_dados_cliente.R
\name{Banco_dados_cliente}
\alias{Banco_dados_cliente}
\title{Banco_dados_cliente}
\usage{
Banco_dados_cliente(
  dados,
  vars_quero,
  salvar = NULL,
  esp = FALSE,
  filtro_titulo = TRUE,
  freeze_titulo = TRUE,
  borda_e_cor_titulo = TRUE,
  cor_titulo_back = "#808080",
  cor_titulo_fonte = "white"
)
}
\arguments{
\item{dados}{Um `data.frame` contendo os dados no formato que vem do servidor.
Deve incluir todas as variáveis disponíveis para seleção}

\item{vars_quero}{Vetor de strings com os nomes das variáveis que serão
selecionadas e organizadas na planilha. A ordem das variáveis neste vetor
será mantida no arquivo final.}

\item{salvar}{Caminho para salvar o arquivo Excel gerado. Se `NULL`, a função
retorna o objeto `Workbook` sem salvar o arquivo. O padrão é `NULL`}

\item{esp}{Lógico. Indica se o nome da aba da planilha será em espanhol. Se
`TRUE`, a aba será chamada "Base de Datos"; caso contrário, será
"Banco de Dados". O padrão é `FALSE`}

\item{filtro_titulo}{Lógico. Determina se será adicionado um filtro na primeira
linha (título das colunas). O padrão é `TRUE`}

\item{freeze_titulo}{Lógico. Indica se a primeira linha da planilha (títulos
das colunas) será congelada, permitindo que permaneça visível ao rolar a
planilha para baixo. O padrão é `TRUE`}

\item{borda_e_cor_titulo}{Lógico. Define se será aplicada uma borda e cor de
fundo no título das colunas. O padrão é `TRUE`}

\item{cor_titulo_back}{String. Define a cor de fundo aplicada ao título das
colunas. O padrão é `"#808080"`}

\item{cor_titulo_fonte}{String. Define a cor da fonte aplicada ao título das
colunas. O padrão é `"white"`}
}
\value{
Retorna um objeto `Workbook` se `salvar` for `NULL`. Caso contrário,
salva o arquivo Excel no caminho especificado.
}
\description{
Esta função gera um arquivo Excel formatado a partir de um
conjunto de dados fornecido, com as variáveis organizadas e estilizadas
conforme as opções fornecidas. Permite a inclusão de filtros, congelamento
de títulos, e estilização do cabeçalho (com bordas e cores personalizadas).

O arquivo pode ser salvo em um local especificado ou o objeto `Workbook` pode
 ser retornado.
}
\details{
Consute o livro para mais detalhes e exemplos.
}
\examples{

# Abrir o Workbook gerado em 'Dicionario_cliente'
openxlsx::openXL(
  file = IPpackage::Dicionario_cliente(
    dicionario = IPpackage::IPpackage_dicionario,
    vars_quero = c("id","v2", "v1", "v19_2", "v19mrg", "v10_2"),
    esp = TRUE,
    coluna_obs = FALSE,
    filtro_titulo = TRUE,
    freeze_titulo = TRUE,
    cor_titulo_back = "#808080",
    cor_titulo_fonte ="white"
  )
)

# Abrir o Workbook gerado em 'Banco_dados_cliente'
openxlsx::openXL(
  file = IPpackage::Banco_dados_cliente(
    dados = IPpackage::IPpackage_exemplo,
    vars_quero = c("id","v2", "v1", "v19_2", "v10_2"),
    salvar = NULL,
    esp = TRUE,
    filtro_titulo = TRUE,
    freeze_titulo = TRUE,
    cor_titulo_back = "#808080",
    cor_titulo_fonte ="white"
  )
)

}
