% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consistencia.R
\name{consistencia}
\alias{consistencia}
\title{consistencia}
\usage{
consistencia(
  log_consistenia = NULL,
  x,
  vars,
  nome = NULL,
  regra,
  pode_falta = FALSE,
  show = FALSE
)
}
\arguments{
\item{log_consistenia}{log_consistenia}

\item{x}{x}

\item{vars}{vars}

\item{nome}{nome}

\item{regra}{regra}

\item{pode_falta}{pode_falta}

\item{show}{show}
}
\description{
x
}
\details{
Consute o livro para mais detalhes e exemplos.
}
\examples{

data=IPpackage::IPpackage_exemplo
##Tudo OK
log_consistenia=consistencia(log_consistenia=NULL,x=data,vars="v1",regra="100",
pode_falta=FALSE,show=c("id","v1"))
log_consistenia=consistencia(log_consistenia=log_consistenia,x=data,vars="v2",
regra="v1\%nin\%c(1)",pode_falta=FALSE,show=c("id","v1","v2"))
log_consistenia=consistencia(log_consistenia=log_consistenia,x=data,
vars=paste0("v3_",1:3),regra="v1\%nin\%c(1)",pode_falta=TRUE,show=c("id","v1","v2")
,nome="G003")
log_consistenia=consistencia(log_consistenia=log_consistenia,x=data,
vars=paste0("v",4:11),regra="100",pode_falta=TRUE,show=c("id","v1","v2"))
##Com Erro
data$v1[3]=NA;data$v1[4]<-1
log_consistenia=consistencia(log_consistenia=NULL,x=data,vars="v1",regra="100",
pode_falta=FALSE,show=c("id","v1","v2"))
log_consistenia=consistencia(log_consistenia=log_consistenia,x=data,vars="v2",
regra="v1\%nin\%c(1)",pode_falta=FALSE,show=c("id","v1","v2"))
log_consistenia=consistencia(log_consistenia=log_consistenia,x=data,
vars=paste0("v3_",1:3),regra="v1\%nin\%c(1)",pode_falta=TRUE,show=c("id","v1",
paste0("v3_",1:3)),nome="G003")

#bbbb
}
