% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consistencia_abertas.R
\name{consistencia_abertas}
\alias{consistencia_abertas}
\title{consistencia_abertas}
\usage{
consistencia_abertas(
  log_consistencia_aberta = NULL,
  xdf,
  var_id,
  var_codificada,
  var_orig = NA,
  show = FALSE,
  obs = NA
)
}
\arguments{
\item{log_consistencia_aberta}{Se a variável em questão for a primeira a ser testada,
'log_consistencia_aberta' é definida como 'NULL'; no entanto, se um arquivo no formato de
saída dessa função já existir, 'log_consistencia_aberta' será o seu data.frame correspondente}

\item{xdf}{Banco de dados}

\item{var_id}{Variável no Banco de Dados com o 'id' do entrevistado}

\item{var_codificada}{Variável com a codificação da 'var_orig'}

\item{var_orig}{Se 'var_orig' for igual a NULL, a função concatenará 'var_codificada'
com '_orig'. Se a variável original não for essa concatenação, por favor, especifique-a aqui}

\item{show}{'TRUE' para mostrar 'var_id', 'var_codificada' e 'var_orig' em caso de erro -serão
exibidas apenas as variáveis indicadas e as linhas com erros; 'FALSE' para não
imprimir as linhas, mesmo na presença de erros.}

\item{obs}{Colocar essa observação em 'descricao'.}
}
\description{
A função foi criada com o propósito de verificar se todos os textos
presentes em 'var_orig' foram codificados na 'var_codificada'.
}
\details{
Consute o livro para mais detalhes e exemplos.
}
\examples{

data=IPpackage::IPpackage_exemplo

##Tudo OK
log_consistencia_aberta=IPpackage::consistencia_abertas(
  log_consistencia_aberta=NULL,xdf=data,var_id="id",var_codificada="v14",
  var_orig=NA,show=FALSE,obs=NA)

##ERRO
data$v14[3]=NA
log_consistencia_aberta=IPpackage::consistencia_abertas(
  log_consistencia_aberta=log_consistencia_aberta,xdf=data,var_id="id",var_codificada="v14",
  var_orig=NA,show=FALSE,obs=NA)

log_consistencia_aberta=IPpackage::consistencia_abertas(
  log_consistencia_aberta=log_consistencia_aberta,xdf=data,var_id="id",var_codificada="v14",
  var_orig=NA,show=TRUE,obs=NA)

log_consistencia_aberta=IPpackage::consistencia_abertas(
  log_consistencia_aberta=log_consistencia_aberta,xdf=data,var_id="id",var_codificada="v14",
  var_orig=NA,show=TRUE,obs="observação")

}
