% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consistencia_geral.R
\name{consistencia_geral}
\alias{consistencia_geral}
\title{consistencia_geral}
\usage{
consistencia_geral(
  caminho,
  banco_de_dados,
  DADOS_dicionario,
  aba_principal,
  coluna_nome_variavel,
  coluna_variaveis,
  coluna_pode_faltar,
  coluna_regra,
  coluna_pode_faltar_label,
  coluna_e_mrg_ou_nao,
  aba_abertas = NULL,
  coluna_aberta_id = NULL,
  coluna_aberta_codificada = NULL,
  coluna_aberta_original = NULL,
  coluna_aberta_obs = NULL
)
}
\arguments{
\item{caminho}{caminho completo do Excel - de 'C:/' à '.xlsx'}

\item{banco_de_dados}{Banco de dados}

\item{DADOS_dicionario}{Dicionário com as colunas 'opcao_cod', 'opcao_label',
'opcao_variavel' e 'pergunta_enunciado'}

\item{aba_principal}{Nome da aba/sheet que contém informações cruciais, como
regras, variáveis e nomes das variáveis}

\item{coluna_nome_variavel}{Nome da coluna em 'aba_principal' que contém o nome
das variáveis ou do MRG}

\item{coluna_variaveis}{Nome da coluna em 'aba_principal' que contém a variável
ou variáveis -em caso de MRG}

\item{coluna_pode_faltar}{Nome da coluna em 'aba_principal' que indica a
possibilidade de ausência de respondentes}

\item{coluna_regra}{Nome da coluna em 'aba_principal' que contém informações
sobre a regra da base}

\item{coluna_pode_faltar_label}{Nome da coluna em 'aba_principal' que indica a
possibilidade de ausência ou presença de labels}

\item{coluna_e_mrg_ou_nao}{Nome da coluna em 'aba_principal' que informa se se
trata ou não de um MRG}

\item{aba_abertas}{Nome da aba/sheet que contém variáveis abertas - 'NULL' caso
não queira analisar as abertas}

\item{coluna_aberta_id}{Nome da coluna em 'aba_abertas' que indica qual coluna
em 'banco_de_dados' possui o 'id' - 'NULL' caso não queira analisar as abertas}

\item{coluna_aberta_codificada}{Nome da coluna em 'aba_abertas' que contém
informações sobre a variável codificada - 'NULL' caso não queira analisar as abertas}

\item{coluna_aberta_original}{Nome da coluna em 'aba_abertas' que contém
informações sobre a variável original - 'NULL' caso não queira analisar as abertas}

\item{coluna_aberta_obs}{Nome da coluna em 'aba_abertas' que indica se deseja
ou não adicionar uma observação à consistência das variáveis abertas - 'NULL'
caso não queira analisar as abertas}
}
\description{
Aprimoramos esse processo ao consolidar todas as informações
relevantes sobre a consistência em uma planilha do Excel, que posteriormente
utilizamos como referência para executar a verificação de consistência por meio
desta função.
}
\details{
Consute o livro para mais detalhes e exemplos.
}
\examples{

# log_consistenia=IPpackage::consistencia_geral(
#   caminho="C:/i/Innovare Pesquisa/Projetos - Documentos/2023.009.00 - CAIXA
#IMAGEM 2023/Estatística/Processamentos/Auxílio finalização/2023.009.02 - Caixa
#Imagem 2023 - Labels.xlsx"
#   ,banco_de_dados=base::readRDS(file = "C:/SSD Externo/Work/InnovarePesquisa/
#Projects/43---2023.009.00 - CAIXA IMAGEM 2023/00.Drive/Banco de dados/02.final.rds")
#   ,DADOS_dicionario=openxlsx::read.xlsx("C:/SSD Externo/Work/InnovarePesquisa/
#Projects/43---2023.009.00 - CAIXA IMAGEM 2023/Output/caixa_imagem_2023_Dicionário.xlsx")
#   ,aba_principal="Labels perguntas"
#   ,coluna_nome_variavel="Variável"
#   ,coluna_variaveis="vars"
#   ,coluna_pode_faltar="pd_falt_cons"
#   ,coluna_regra="C_regra"
#   ,coluna_pode_faltar_label="pd_falt_labl"
#   ,coluna_e_mrg_ou_nao="obs"
#   ,aba_abertas=NULL
#   ,coluna_aberta_id=NULL
#   ,coluna_aberta_codificada=NULL
#   ,coluna_aberta_original=NULL
#   ,coluna_aberta_obs=NULL
# )

}
