% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consistencia_label.R
\name{consistencia_label}
\alias{consistencia_label}
\title{consistencia_label}
\usage{
consistencia_label(
  log_consistenia_label = NULL,
  xdf,
  xdicionario,
  vars,
  var_mrg_nome = NA,
  mrg_citou = FALSE,
  pode_falta = FALSE,
  show = FALSE
)
}
\arguments{
\item{log_consistenia_label}{Se a variável em questão for a primeira a ser testada,
'log_consistenia_label' é definida como 'NULL'; no entanto, se um arquivo no formato de
saída dessa função já existir, 'log_consistenia_label' será o seu data.frame correspondente}

\item{xdf}{Banco de dados}

\item{xdicionario}{Dicionário com as colunas 'opcao_cod', 'opcao_label',
'opcao_variavel', 'pergunta_enunciado'}

\item{vars}{Variável ou, no caso de um MRG, vetor de variáveis}

\item{var_mrg_nome}{'NA' caso não se trate de um MRG; caso contrário, nome do MRG}

\item{mrg_citou}{'TRUE' para indicar se tratar de MRG do tipo Citou;
caso contrário, 'FALSE'.}

\item{pode_falta}{'TRUE' para indicar a possibilidade de labels ausentes;
'FALSE' para indicar a ausência de tolerância para labels não fornecidas.}

\item{show}{'TRUE' para imprimir a frequência com Label e Enunciado;
'FALSE', caso contrário.}
}
\description{
A função foi desenvolvida para avaliar se todas as variáveis possuem
 Label e Enunciado no Dicionário e se existe algum código repetido.
}
\details{
Consute o livro para mais detalhes e exemplos.
}
\examples{

#Sem exemplo
}
