% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dicionario_cliente.R
\name{Dicionario_cliente}
\alias{Dicionario_cliente}
\title{Dicionario_cliente}
\usage{
Dicionario_cliente(
  dicionario,
  vars_quero,
  salvar = NULL,
  esp = FALSE,
  coluna_obs = FALSE,
  filtro_titulo = TRUE,
  freeze_titulo = TRUE,
  cor_titulo_back = "#808080",
  cor_titulo_fonte = "white"
)
}
\arguments{
\item{dicionario}{Um data.frame contendo o dicionário base. Deve conter as
colunas "opcao_variavel", "pergunta_enunciado", "opcao_cod", e "opcao_label",
além de uma coluna opcional de observações ("Obs") caso coluna_obs seja TRUE}

\item{vars_quero}{Um vetor de caracteres especificando as variáveis de
interesse na ordem que elas devem aparecer no dicionário}

\item{salvar}{Um caminho para salvar o arquivo Excel gerado. Se NULL, a
função retorna o objeto Workbook sem salvar o arquivo}

\item{esp}{Um valor lógico. Se TRUE, o dicionário será gerado em espanhol.
Se FALSE, o idioma padrão será o português}

\item{coluna_obs}{Um valor lógico. Indica se a coluna de observações ("Obs")
deve ser incluída no dicionário. Se TRUE, a coluna será incluída}

\item{filtro_titulo}{Um valor lógico. Se TRUE, adiciona um filtro na
primeira linha do dicionário (título das colunas)}

\item{freeze_titulo}{Um valor lógico. Se TRUE, a primeira linha do
dicionário (título das colunas) será congelada ao rolar a planilha}

\item{cor_titulo_back}{Cor de fundo para o título do dicionário (primeira
linha). O padrão é "#808080" (cinza)}

\item{cor_titulo_fonte}{Cor da fonte (texto) do título do dicionário
(primeira linha). O padrão é "white" (branco)}
}
\value{
Retorna um objeto Workbook se o parâmetro salvar for NULL.
Caso contrário, salva o arquivo Excel no local especificado.
}
\description{
Esta função gera um dicionário de dados personalizado a partir de
um dicionário (no formato do que entra na função de processamento).

O dicionário resultante contém variáveis específicas (definidas em vars_quero)
organizadas e formatadas de acordo com as opções fornecidas, como a inclusão de
filtros, congelamento de títulos e estilização visual.

Pode gerar uma planilha Excel formatada com as variáveis e descrições desejadas,
com opção de salvar o arquivo ou retornar o objeto Workbook.
}
\details{
Consute o livro para mais detalhes e exemplos.
}
\examples{

# Abrir o Workbook gerado em 'Dicionario_cliente'
openxlsx::openXL(
  file = IPpackage::Dicionario_cliente(
    dicionario = IPpackage::IPpackage_dicionario,
    vars_quero = c("id","v2", "v1", "v19_2", "v19mrg", "v10_2"),
    esp = TRUE,
    coluna_obs = FALSE,
    filtro_titulo = TRUE,
    freeze_titulo = TRUE,
    cor_titulo_back = "#808080",
    cor_titulo_fonte ="white"
  )
)

# Abrir o Workbook gerado em 'Banco_dados_cliente'
openxlsx::openXL(
  file = IPpackage::Banco_dados_cliente(
    dados = IPpackage::IPpackage_exemplo,
    vars_quero = c("id","v2", "v1", "v19_2", "v10_2"),
    salvar = NULL,
    esp = TRUE,
    filtro_titulo = TRUE,
    freeze_titulo = TRUE,
    cor_titulo_back = "#808080",
    cor_titulo_fonte ="white"
  )
)

}
