% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Felipe_radar.R
\name{Felipe_radar}
\alias{Felipe_radar}
\title{Felipe_radar}
\usage{
Felipe_radar(
  plot.data,
  base.size = 15,
  font.radar = "sans",
  values.radar = c("0\%", "10\%", "20\%", "30\%", "40\%", "50\%", "60\%", "70\%",
    "80\%", "90\%", "100\%"),
  axis.labels = colnames(plot.data)[-1],
  grid.min = 0,
  grid.um = 0.1,
  grid.dois = 0.2,
  grid.tres = 0.3,
  grid.quatro = 0.4,
  grid.cinco = 0.5,
  grid.seis = 0.6,
  grid.sete = 0.7,
  grid.oito = 0.8,
  grid.nove = 0.9,
  grid.dez = 1,
  centre.y = grid.min - ((1/9) * (grid.dez - grid.min)),
  plot.extent.x.sf = 1,
  plot.extent.y.sf = 1.2,
  x.centre.range = 0.02 * (grid.dez - centre.y),
  label.centre.y = FALSE,
  grid.line.width = 0.5,
  gridline.min.linetype = "longdash",
  gridline.um.linetype = "longdash",
  gridline.dois.linetype = "longdash",
  gridline.tres.linetype = "longdash",
  gridline.quatro.linetype = "longdash",
  gridline.cinco.linetype = "longdash",
  gridline.seis.linetype = "longdash",
  gridline.sete.linetype = "longdash",
  gridline.oito.linetype = "longdash",
  gridline.nove.linetype = "longdash",
  gridline.dez.linetype = "longdash",
  gridline.min.colour = "grey",
  gridline.um.colour = "grey",
  gridline.dois.colour = "grey",
  gridline.tres.colour = "grey",
  gridline.quatro.colour = "grey",
  gridline.cinco.colour = "grey",
  gridline.seis.colour = "grey",
  gridline.sete.colour = "grey",
  gridline.oito.colour = "grey",
  gridline.nove.colour = "grey",
  gridline.dez.colour = "grey",
  grid.label.size = 4,
  gridline.label.offset = -0.1 * (grid.dez - centre.y),
  label.gridline.min = TRUE,
  label.gridline.um = TRUE,
  label.gridline.dois = TRUE,
  label.gridline.tres = TRUE,
  label.gridline.quatro = TRUE,
  label.gridline.cinco = TRUE,
  label.gridline.seis = TRUE,
  label.gridline.sete = TRUE,
  label.gridline.oito = TRUE,
  label.gridline.nove = TRUE,
  label.gridline.dez = TRUE,
  axis.label.offset = 1.15,
  axis.label.size = 5,
  axis.line.colour = "grey",
  group.line.width = 1,
  group.point.size = 3,
  group.colours = NULL,
  background.circle.colour = "#D7D6D1",
  background.circle.transparency = 0.2,
  plot.legend = ifelse(nrow(plot.data) > 1, TRUE, FALSE),
  legend.title = "",
  plot.title = "",
  plot.title.hjust = 0.5,
  plot.title.size = 15,
  legend.text.size = 14,
  legend.position = "top",
  fill = FALSE,
  fill.alpha = 0.5,
  graf.limits = c(2, 2)
)
}
\arguments{
\item{plot.data}{plot.data}

\item{base.size}{= 15}

\item{font.radar}{= "sans"}

\item{values.radar}{= values.radar}

\item{axis.labels}{= colnames(plot.data)[-1]}

\item{grid.min}{= 0}

\item{grid.um}{= 0.10}

\item{grid.dois}{= 0.20}

\item{grid.tres}{= 0.30}

\item{grid.quatro}{= 0.40}

\item{grid.cinco}{= 0.50}

\item{grid.seis}{= 0.60}

\item{grid.sete}{= 0.70}

\item{grid.oito}{= 0.80}

\item{grid.nove}{= 0.90}

\item{grid.dez}{= 1}

\item{centre.y}{= grid.min - ((1/9) *(grid.dez - grid.min))}

\item{plot.extent.x.sf}{= 1}

\item{plot.extent.y.sf}{= 1.2}

\item{x.centre.range}{= 0.02 * (grid.dez - centre.y)}

\item{label.centre.y}{= FALSE}

\item{grid.line.width}{= 0.5}

\item{gridline.min.linetype}{= "longdash"}

\item{gridline.um.linetype}{= "longdash"}

\item{gridline.dois.linetype}{= "longdash"}

\item{gridline.tres.linetype}{= "longdash"}

\item{gridline.quatro.linetype}{= "longdash"}

\item{gridline.cinco.linetype}{= "longdash"}

\item{gridline.seis.linetype}{= "longdash"}

\item{gridline.sete.linetype}{= "longdash"}

\item{gridline.oito.linetype}{= "longdash"}

\item{gridline.nove.linetype}{= "longdash"}

\item{gridline.dez.linetype}{= "longdash"}

\item{gridline.min.colour}{= "grey"}

\item{gridline.um.colour}{= "grey"}

\item{gridline.dois.colour}{= "grey"}

\item{gridline.tres.colour}{= "grey"}

\item{gridline.quatro.colour}{= "grey"}

\item{gridline.cinco.colour}{= "grey"}

\item{gridline.seis.colour}{= "grey"}

\item{gridline.sete.colour}{= "grey"}

\item{gridline.oito.colour}{= "grey"}

\item{gridline.nove.colour}{= "grey"}

\item{gridline.dez.colour}{= "grey"}

\item{grid.label.size}{= 4}

\item{gridline.label.offset}{= -0.1 * (grid.dez - centre.y)}

\item{label.gridline.min}{= TRUE}

\item{label.gridline.um}{= TRUE}

\item{label.gridline.dois}{= TRUE}

\item{label.gridline.tres}{= TRUE}

\item{label.gridline.quatro}{= TRUE}

\item{label.gridline.cinco}{= TRUE}

\item{label.gridline.seis}{= TRUE}

\item{label.gridline.sete}{= TRUE}

\item{label.gridline.oito}{= TRUE}

\item{label.gridline.nove}{= TRUE}

\item{label.gridline.dez}{= TRUE}

\item{axis.label.offset}{= 1.15}

\item{axis.label.size}{= 5}

\item{axis.line.colour}{= "grey"}

\item{group.line.width}{= 1}

\item{group.point.size}{= 3}

\item{group.colours}{= NULL}

\item{background.circle.colour}{= "#D7D6D1"}

\item{background.circle.transparency}{= 0.2}

\item{plot.legend}{=ifelse(nrow(plot.data) >1,TRUE,FALSE)}

\item{legend.title}{= ""}

\item{plot.title}{= ""}

\item{plot.title.hjust}{=.5}

\item{plot.title.size}{=15}

\item{legend.text.size}{= 14}

\item{legend.position}{= "top"}

\item{fill}{= FALSE}

\item{fill.alpha}{= 0.5}

\item{graf.limits}{=c(2,2)}
}
\description{
x
}
\details{
Consute o livro para mais detalhes e exemplos.
}
\examples{
library(dplyr)
library(scales)
library(tibble)

mtcars_radar <- mtcars \%>\%
  tibble::as_tibble(rownames = "group") \%>\%
  dplyr::mutate_at(vars(-group), scales::rescale) \%>\%
  utils::tail(4) \%>\%
  dplyr::select(1:10)
Felipe_radar(mtcars_radar,group.colours=c("#FFB400","#FF5A5F", "#007A87","#8CE071"),
values.radar = c("0\%","","20\%","","40\%","","60\%","","80\%","","100\%"),grid.label.size=5,
background.circle.colour="white",axis.label.size = 5)

##Comparando com ggradar
#library(ggradar)
#ggradar(mtcars_radar)
##Remodeno o fundo e salvando
#a=Felipe_radar(mtcars_radar,group.colours=c("#FFB400","#FF5A5F", "#007A87","#8CE071"),
#values.radar = c("0\%","","20\%","","40\%","","60\%","","80\%","","100\%"),grid.label.size=5,
#background.circle.colour="white",axis.label.size = 5,fill=T);a
#ggsave(a, file="Felipe_radar.png" , width=13, height=7, bg = "transparent")
#library(magick)
#magick::image_write(magick::image_transparent(magick::image_read("Felipe_radar.png"),
#white'),"Felipe_radar.png")



}
