% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUN_Citou.R
\name{FUN_Citou}
\alias{FUN_Citou}
\title{FUN_Citou}
\usage{
FUN_Citou(TABELA, DICIONARIO, lista_variaveis, adc_labels = TRUE)
}
\arguments{
\item{TABELA}{Banco de dados a ser analisado.}

\item{DICIONARIO}{Dicionário correspondente à TABELA.}

\item{lista_variaveis}{Variáveis que serão pivotadas (vetor).}

\item{adc_labels}{Se adc_labels=TRUE, adicionar uma coluna com o label.}
}
\description{
Pivota e depois calcula a frequência simples (ponderada e não ponderada)
 das variáveis indicadas em 'lista_variaveis' (MRG).
 Código 1 = Citou; código diferente de 1 = não citou.
}
\details{
Consute o livro para mais detalhes e exemplos.
}
\examples{

#Passando 'outros' (v10) para do tipo citou/não citou
library(dplyr)
library(tidyr)
TABELA = IPpackage::IPpackage_exemplo \%>\%
 dplyr::mutate(
   v10_o = dplyr::case_when(
     is.na(v4) & is.na(v10) ~ NA,
     is.na(v10) ~ 2,
     !is.na(v10) ~ 1
   )
 )

DICIONARIO = IPpackage::IPpackage_dicionario \%>\%
 rbind(
   tibble::tibble(
     opcao_cod = c(1, 2),
     opcao_label = c("Citou", "Não citou"),
     opcao_variavel = rep("v10_o", 2),
     pergunta_enunciado = rep("v10 - Citou (Outros)", 2)
   )
 )

#MRG sem erro
FUN_Citou(
 TABELA = TABELA,
 DICIONARIO = DICIONARIO,
 lista_variaveis = list("v3mrg" = c("v4","v5","v6","v7","v8","v9","v10_o","v11")),
 adc_labels = TRUE
)
#MRG com erro:
#variável que tem código diferente de 1 e/ou 2
#enunciado de alguma variável faltando
FUN_Citou(
 TABELA = TABELA,
 DICIONARIO = DICIONARIO\%>\%
   dplyr::mutate(
     pergunta_enunciado=ifelse(opcao_variavel \%in\% c("v5","v11"),NA,pergunta_enunciado)
   ),
 lista_variaveis = list("v3mrg" = c("v4","v5","v6","v7","v8","v9","v10","v11")),
 adc_labels = TRUE
)


}
