% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUN_cruzaCitou.R
\name{FUN_cruzaCitou}
\alias{FUN_cruzaCitou}
\title{FUN_cruzaCitou}
\usage{
FUN_cruzaCitou(TABELA, DICIONARIO, var1, varCitou, var_fecha)
}
\arguments{
\item{TABELA}{Banco de dados a ser analisado.}

\item{DICIONARIO}{Dicionário correspondente à TABELA.}

\item{var1}{Variável que será cruzada pela 'varCitou'.}

\item{varCitou}{Variáveis do tipo citou/não citou.}

\item{var_fecha}{Variável na coluna.}
}
\description{
Pivota e depois calcula a frequência simples (ponderada e não ponderada)
 das variáveis indicadas em 'varCitou' (Código 1 = Citou; código diferente de 1 = não citou)
 cruzada pela variável indicada em 'ind'.
 O parâmetro 'var_fecha' define a variável na coluna.
}
\details{
Consute o livro para mais detalhes e exemplos.
}
\examples{

library(IPpackage)
library(dplyr)
library(tidyr)
library(tibble)
library(stringr)
library(purrr)
#Passando 'outros' (v10) para do tipo citou/não citou
TABELA = IPpackage::IPpackage_exemplo \%>\%
 dplyr::mutate(
   v10_1_o = dplyr::case_when(
     base::is.na(v4) & base::is.na(v10_1) ~ NA,
     base::is.na(v10_1) ~ 2,
     !base::is.na(v10_1) ~ 1
   ),
   v10_2_o = dplyr::case_when(
     base::is.na(v4) & base::is.na(v10_2) ~ NA,
     base::is.na(v10_2) ~ 2,
     !base::is.na(v10_2) ~ 1
   ),
   v10_outros = dplyr::case_when(
     base::is.na(v10_1_o) & base::is.na(v10_2_o) ~ NA,
     v10_1_o == 1 | v10_2_o == 1 ~ 1,
     TRUE ~ 2
   )
 )
TABELA \%>\% dplyr::count(v4,v10_1,v10_1_o)
TABELA \%>\% dplyr::count(v4,v10_2,v10_2_o)
TABELA \%>\% dplyr::count(v4,v10_1_o,v10_2_o,v10_outros)

DICIONARIO = IPpackage::IPpackage_dicionario \%>\%
 base::rbind(
   tibble::tibble(
     opcao_cod = base::rep(c(1, 2),3),
     opcao_label = base::rep(c("Citou", "Não citou"),3),
     opcao_variavel = c(
       base::rep("v10_outros", 2),
       base::rep("v10_1_o", 2),
       base::rep("v10_2_o", 2)
     ),
     pergunta_enunciado = c(
       base::rep("v10_outros - Citou (Outros)", 2),
       base::rep("v10_1_o - Citou (Outros)", 2),
       base::rep("v10_2_o - Citou (Outros)", 2)
     )
   )
 )


#Rodando a função-------------------------------------------------------------

#1. Categoria MRG (v15mrg - v15, v16) vs Citou MRG (v4mrg - v4 a v11):----------

##Fechando na v15mrg (Divide pela base - quem falou pelo menos um cód em v15mrg)
FUN_cruzaCitou(
 TABELA = TABELA,
 DICIONARIO = DICIONARIO,
 var1 = base::list(
   "v15mrg" = c("v15", "v16")
 ),
 varCitou = base::list(
   "v4mrg" = c("v4", "v5", "v6", "v7", "v8", "v9", "v10_outros", "v11")
 ),
 var_fecha = "v15mrg"
) \%>\%
 base::data.frame()

##Fechando na v4mrg (Divide pela soma de n - soma 100\%)
FUN_cruzaCitou(
 TABELA = TABELA,
 DICIONARIO = DICIONARIO,
 var1 = base::list(
   "v15mrg" = c("v15", "v16")
 ),
 varCitou = base::list(
   "v4mrg" = c("v4", "v5", "v6", "v7", "v8", "v9", "v10_outros", "v11")
 ),
 var_fecha = "v4mrg"
) \%>\%
 base::data.frame()

#2.Categoria Isolada (v15) vs Citou MRG (v4mrg - v4 a v11):---------------------

##Fechando na v15 (Divide pela base - quem falou cód em v15)
FUN_cruzaCitou(
 TABELA = TABELA,
 DICIONARIO = DICIONARIO,
 var1 = "v15",
 varCitou = base::list(
   "v4mrg" = c("v4", "v5", "v6", "v7", "v8", "v9", "v10_outros", "v11")
 ),
 var_fecha = "v15"
) \%>\%
 base::data.frame()

##Fechando na v4mrg (Divide pela soma de n - soma 100\%)
FUN_cruzaCitou(
 TABELA = TABELA,
 DICIONARIO = DICIONARIO,
 var1 = "v15",
 varCitou = base::list(
   "v4mrg" = c("v4", "v5", "v6", "v7", "v8", "v9", "v10_outros", "v11")
 ),
 var_fecha = "v4mrg"
) \%>\%
 base::data.frame()

#3.Categoria MRG (v15mrg - v15, v16) vs Citou Isolada v4:-----------------------

##Fechando na v15mrg (Divide pela base - quem falou pelo menos um cód em v15mrg)
FUN_cruzaCitou(
 TABELA = TABELA,
 DICIONARIO = DICIONARIO,
 var1 = base::list("v15mrg" = c("v15", "v16")),
 varCitou = base::list("v4" = "v4"),
 var_fecha = "v15mrg"
) \%>\%
 base::data.frame()

##Fechando na v4 (Divide pela soma de n - soma 100\%)
FUN_cruzaCitou(
 TABELA = TABELA,
 DICIONARIO = DICIONARIO,
 var1 = base::list("v15mrg" = c("v15", "v16")),
 varCitou = base::list("v4" = "v4"),
 var_fecha = "v4"
) \%>\%
 base::data.frame()

#4.Categoria Isolada v15 vs Citou Isolada v4:-----------------------------------

##Fechando na v15 (Divide pela base - quem falou cód em v15)
FUN_cruzaCitou(
 TABELA = TABELA,
 DICIONARIO = DICIONARIO,
 var1 = "v15",
 varCitou = base::list("v4" = "v4"),
 var_fecha = "v15"
) \%>\%
 base::data.frame()

## Fechando na v4 (Divide pela soma de n - soma 100\%)
FUN_cruzaCitou(
 TABELA = TABELA,
 DICIONARIO = DICIONARIO,
 var1 = "v15",
 varCitou = base::list("v4" = "v4"),
 var_fecha = "v4"
) \%>\%
 base::data.frame()

}
