% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUN_cruzaMedia.R
\name{FUN_cruzaMedia}
\alias{FUN_cruzaMedia}
\title{FUN_cruzaMedia}
\usage{
FUN_cruzaMedia(TABELA, DICIONARIO, var1, ind)
}
\arguments{
\item{TABELA}{Banco de dados a ser analisado.}

\item{DICIONARIO}{Dicionário correspondente à TABELA.}

\item{var1}{Variável categórica (pode ser lista - MRG).}

\item{ind}{Variável numérica (pode ser lista - MRG).}
}
\description{
Pivota e depois calcula a média (ponderada e não ponderada) da
variável indicada em 'ind' cruzada pela variável indicada em 'var1'.
}
\details{
Consute o livro para mais detalhes e exemplos.
}
\examples{

library(dplyr)
library(tidyr)
library(purrr)
library(IPpackage)

#Rodando a função-------------------------------------------------------------

# Categoria MRG (v15mrg - v15, v16) vs Média MRG (v17mrg - v17, v18):
IPpackage::FUN_cruzaMedia(
 TABELA = IPpackage::IPpackage_exemplo,
 DICIONARIO = IPpackage::IPpackage_dicionario,
 var1 = base::list("v15mrg" = c("v15", "v16")),
 ind = base::list("v17mrg" = c("v17", "v18"))
) \%>\%
 base::data.frame()

# Categoria Isolada (v15) vs Média MRG (v17mrg - v17, v18):
IPpackage::FUN_cruzaMedia(
 TABELA = IPpackage::IPpackage_exemplo,
 DICIONARIO = IPpackage::IPpackage_dicionario,
 var1 = "v15",
 ind = base::list("v17mrg" = c("v17", "v18"))
) \%>\%
 base::data.frame()

# Categoria MRG (v15mrg - v15, v16) vs Média Isolada (v17):
IPpackage::FUN_cruzaMedia(
 TABELA = IPpackage::IPpackage_exemplo,
 DICIONARIO = IPpackage::IPpackage_dicionario,
 var1 = base::list("v15mrg" = c("v15", "v16")),
 ind = "v17"
) \%>\%
 base::data.frame()

# Categoria Isolada (v16) vs Média Isolada (v17):
IPpackage::FUN_cruzaMedia(
 TABELA = IPpackage::IPpackage_exemplo,
 DICIONARIO = IPpackage::IPpackage_dicionario,
 var1 = "v16",
 ind = "v17"
) \%>\%
 base::data.frame()

#Warnings---------------------------------------------------------------------

# warning1 (ao forçar virar numérica, algo virou NA)
IPpackage::FUN_cruzaMedia(
 TABELA = IPpackage::IPpackage_exemplo \%>\%
   dplyr::mutate(v17 = ifelse(id == 4, "texto", v17)),
 DICIONARIO = IPpackage::IPpackage_dicionario,
 var1 = "v16",
 ind = "v17"
)

# warning2 (converteu em númerico sem problemas)
IPpackage::FUN_cruzaMedia(
 TABELA = IPpackage::IPpackage_exemplo \%>\%
   dplyr::mutate(v17 = as.character(v17)),
 DICIONARIO = IPpackage::IPpackage_dicionario,
 var1 = "v16",
 ind = "v17"
)

# warning3 (se a variável só tiver NA)
IPpackage::FUN_cruzaMedia(
 TABELA = IPpackage::IPpackage_exemplo \%>\%
   dplyr::mutate(v17 = NA),
 DICIONARIO = IPpackage::IPpackage_dicionario,
 var1 = "v16",
 ind = "v17"
)

}
