% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUN_cruzaVar.R
\name{FUN_cruzaVar}
\alias{FUN_cruzaVar}
\title{FUN_cruzaVar}
\usage{
FUN_cruzaVar(TABELA, DICIONARIO, var1, var2, var_fecha)
}
\arguments{
\item{TABELA}{Banco de dados a ser analisado.}

\item{DICIONARIO}{Dicionário correspondente à TABELA.}

\item{var1}{Primeira variável do cruzamento.}

\item{var2}{Segunda variável do cruzamento.}

\item{var_fecha}{Variável na coluna.}
}
\description{
Calcula a frequência cruzada entre duas variáveis (ponderada e não ponderada).
O parâmetro 'var_fecha' define a variável na coluna.
}
\details{
Consute o livro para mais detalhes e exemplos.
}
\examples{

library(dplyr)
library(IPpackage)

# Variável na coluna: v1
IPpackage::FUN_cruzaVar(
 TABELA = IPpackage::IPpackage_exemplo,
 DICIONARIO = IPpackage::IPpackage_dicionario,
 var1 = "v1",
 var2 = "v10",
 var_fecha = "v1"
)

# Variável na coluna: v10
IPpackage::FUN_cruzaVar(
 TABELA = IPpackage::IPpackage_exemplo,
 DICIONARIO = IPpackage::IPpackage_dicionario,
 var1 = "v1",
 var2 = "v10",
 var_fecha = "v10"
)


}
