% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUN_isoladas.R
\name{FUN_isoladas}
\alias{FUN_isoladas}
\title{FUN_isoladas}
\usage{
FUN_isoladas(TABELA, DICIONARIO, variaveis, adc_labels = TRUE)
}
\arguments{
\item{TABELA}{Banco de dados a ser analisado.}

\item{DICIONARIO}{Dicionário correspondente à TABELA.}

\item{variaveis}{Variáveis para as quais as frequências devem ser calculadas (vetor).}

\item{adc_labels}{Se adc_labels=TRUE, adicionar uma coluna com o label.}
}
\description{
Determinar a frequência simples das 'variaveis', tanto ponderada quanto não ponderada.
}
\details{
Consute o livro para mais detalhes e exemplos.
}
\examples{

#Uma variável
IPpackage::FUN_isoladas(
  TABELA=IPpackage::IPpackage_exemplo,
  DICIONARIO=IPpackage::IPpackage_dicionario,
  variaveis=c("v1"),
  adc_labels = TRUE
)

#Mais de uma variável
IPpackage::FUN_isoladas(
  TABELA=IPpackage::IPpackage_exemplo,
  DICIONARIO=IPpackage::IPpackage_dicionario,
  variaveis=c("v1","v2"),
  adc_labels = TRUE
)

#Todas as variáveis
IPpackage::FUN_isoladas(
  TABELA=IPpackage::IPpackage_exemplo,
  DICIONARIO=IPpackage::IPpackage_dicionario,
  variaveis=colnames(IPpackage::IPpackage_exemplo),
  adc_labels = TRUE
)

}
