% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUN_media.R
\name{FUN_media}
\alias{FUN_media}
\title{FUN_media}
\usage{
FUN_media(TABELA, lista_variaveis, debug = FALSE)
}
\arguments{
\item{TABELA}{Banco de dados a ser analisado.}

\item{lista_variaveis}{Lista das variáveis que serão pivotadas.}

\item{debug}{Se debug=TRUE, printa cada iteração de variáveis.}
}
\description{
Pivota e depois calcula a média (ponderada e não ponderada) das variáveis indicadas em 'lista_variaveis'.
}
\details{
Consute o livro para mais detalhes e exemplos.
}
\examples{

library(dplyr)
library(tidyr)
library(IPpackage)

# Normal
IPpackage::FUN_media(
 TABELA = IPpackage::IPpackage_exemplo
 ,lista_variaveis = base::list(
   "v2" = "v2",
   "v17mrg" = c("v17", "v18"),
   "v17" = "v17",
   "v18" = "v18"
 )
)

# warning1 (ao forçar virar numérica, algo virou NA)
IPpackage::FUN_media(
 TABELA = IPpackage::IPpackage_exemplo \%>\%
   dplyr::mutate(v17 = base::ifelse(id == 4, "texto", v17))
 ,lista_variaveis = base::list(
   "v2" = "v2",
   "v17mrg" = c("v17", "v18"),
   "v17" = "v17",
   "v18" = "v18"
 )
)

# warning2 (converteu em númerico sem problemas)
IPpackage::FUN_media(
 TABELA = IPpackage::IPpackage_exemplo \%>\%
   dplyr::mutate(v17 = base::as.character(v17))
 ,lista_variaveis = base::list(
   "v2" = "v2",
   "v17mrg" = c("v17", "v18"),
   "v17" = "v17",
   "v18" = "v18"
 )
)

# warning3 (se a variável só tiver NA)
IPpackage::FUN_media(
 TABELA = IPpackage::IPpackage_exemplo \%>\%
   dplyr::mutate(v17 = NA)
 ,lista_variaveis = base::list(
   "v2" = "v2",
   "v17mrg" = c("v17", "v18"),
   "v17" = "v17",
   "v18" = "v18"
 )
)

# Todos os warnings
IPpackage::FUN_media(
 TABELA = IPpackage::IPpackage_exemplo \%>\%
   dplyr::mutate(
     v2 = NA
     ,v17 = base::ifelse(id == 4, "texto", v17)
     ,v18 = base::as.character(v18)
   )
 ,lista_variaveis = base::list(
   "v2" = "v2",
   "v17mrg" = c("v17", "v18"),
   "v17" = "v17",
   "v18" = "v18"
 )
)

}
