% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUN_media_das_medias.R
\name{FUN_media_das_medias}
\alias{FUN_media_das_medias}
\title{FUN_media_das_medias}
\usage{
FUN_media_das_medias(TABELA, lista_variaveis)
}
\arguments{
\item{TABELA}{Banco de dados a ser analisado.}

\item{lista_variaveis}{Lista das variáveis.}
}
\description{
Calcula a média (ponderada e não ponderada) e, se MRG, calcula a média das médias das variáveis indicadas em 'lista_variaveis'.
}
\details{
Consute o livro para mais detalhes e exemplos.
}
\examples{

library(dplyr)
library(tidyr)
library(IPpackage)

# Normal
IPpackage::FUN_media_das_medias(
 TABELA = IPpackage::IPpackage_exemplo
 ,lista_variaveis = list(
   "v2" = "v2",
   "v10mrg" = c("v10", "v11"),
   "v10" = "v10",
   "v11" = "v11"
 )
)

# warning1 (ao forçar virar numérica, algo virou NA)
IPpackage::FUN_media_das_medias(
 TABELA = IPpackage::IPpackage_exemplo \%>\%
   dplyr::mutate(v10 = ifelse(id == 4, "texto", v10))
 ,lista_variaveis = list(
   "v2" = "v2",
   "v10mrg" = c("v10", "v11"),
   "v10" = "v10",
   "v11" = "v11"
 )
)

# warning2 (converteu em númerico sem problemas)
IPpackage::FUN_media_das_medias(
 TABELA = IPpackage::IPpackage_exemplo \%>\%
   dplyr::mutate(v10 = as.character(v10))
 ,lista_variaveis=list(
   "v2" = "v2",
   "v10mrg" = c("v10", "v11"),
   "v10" = "v10",
   "v11" = "v11"
 )
)

# warning3 (se a variável só tiver NA)
IPpackage::FUN_media_das_medias(
 TABELA = IPpackage::IPpackage_exemplo \%>\%
   dplyr::mutate(v10 = NA)
 ,lista_variaveis = list(
   "v2" = "v2",
   "v10mrg" = c("v10", "v11"),
   "v10" = "v10",
   "v11" = "v11"
 )
)

# Todos os warnings
IPpackage::FUN_media_das_medias(
 TABELA = IPpackage::IPpackage_exemplo \%>\%
   dplyr::mutate(
     v2 = NA
     ,v10 = ifelse(id == 4, "texto", v10)
     ,v11 =as.character(v12)
   )
 ,lista_variaveis = list(
   "v2" = "v2",
   "v10mrg" = c("v10", "v11"),
   "v10" = "v10",
   "v11" = "v11"
 )
)

}
