% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUN_MRG.R
\name{FUN_MRG}
\alias{FUN_MRG}
\title{FUN_MRG}
\usage{
FUN_MRG(TABELA, DICIONARIO, lista_variaveis, adc_labels = TRUE)
}
\arguments{
\item{TABELA}{Banco de dados a ser analisado.}

\item{DICIONARIO}{Dicionário correspondente à TABELA.}

\item{lista_variaveis}{Variáveis que serão pivotadas (vetor).}

\item{adc_labels}{Se adc_labels=TRUE, adicionar uma coluna com o label.}
}
\description{
Pivota e depois calcula a frequência simples (ponderada e não ponderada) das variáveis indicadas em 'lista_variaveis' (MRG).
}
\details{
Consute o livro para mais detalhes e exemplos.
}
\examples{

# MRG sem erro
IPpackage::FUN_MRG(
 TABELA=IPpackage::IPpackage_exemplo,
 DICIONARIO=IPpackage::IPpackage_dicionario,
 lista_variaveis=list("v3mrg"=c("v3_1","v3_2","v3_3")),
 adc_labels = TRUE
)

library(dplyr)

# MRG sem a variável no DICIONARIO
IPpackage::FUN_MRG(
 TABELA=IPpackage::IPpackage_exemplo,
 DICIONARIO=IPpackage::IPpackage_dicionario\%>\%dplyr::mutate(opcao_cod=ifelse(
   opcao_variavel=="v3mrg",NA,opcao_cod)),
 lista_variaveis=list("v3mrg"=c("v3_1","v3_2","v3_3")),
 adc_labels = TRUE
)

# MRG com a variável no DICIONARIO, mas faltando Label
IPpackage::FUN_MRG(
 TABELA=IPpackage::IPpackage_exemplo,
 DICIONARIO=IPpackage::IPpackage_dicionario\%>\%
   dplyr::filter(opcao_variavel!="v3mrg"),
 lista_variaveis=list("v3mrg"=c("v3_1","v3_2","v3_3")),
 adc_labels = TRUE
)

}
