% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_freqToDF.R
\name{func_freqToDF}
\alias{func_freqToDF}
\title{func_freqToDF}
\usage{
func_freqToDF(TABELA)
}
\arguments{
\item{TABELA}{Falta escrever.}
}
\description{
Falta escrever.
}
\details{
Consute o livro para mais detalhes e exemplos.
}
\examples{

#Falta escrever.

}
