% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPpackage_dicionario.R
\docType{data}
\name{IPpackage_dicionario}
\alias{IPpackage_dicionario}
\title{Dicionário dos Dados Fictícios}
\format{
`IPpackage_dicionario`
Um data frame com X linhas e 4 colunas:
\describe{
  \item{opcao_cod}{Código.}
  \item{opcao_label}{Label do código.}
  \item{opcao_variavel}{Variável.}
  \item{pergunta_enunciado}{Enunciado.}
}
}
\source{
Fictício
}
\usage{
IPpackage_dicionario
}
\description{
Dicionário dos dados para simular um processamento.
}
\keyword{datasets}
