% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPpackage_exemplo.R
\docType{data}
\name{IPpackage_exemplo}
\alias{IPpackage_exemplo}
\title{World Health Organization TB data}
\format{
## `IPpackage_exemplo`
A data frame with 100 rows and X columns:
\describe{
  \item{id}{ID}
  \item{v1}{Opções}
  \item{v2}{Numérico (só responde quem não respondeu v1==1)}
  \item{v3_1,v3_2,v3_3}{MRG}
  \item{v4,v5,v6,v7,v8,v9,v10,v11}{MRG citou; v10=Outros; v11=NS/NR}
  ...
}
}
\source{
Fictício
}
\usage{
IPpackage_exemplo
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
Report ...
}
\keyword{datasets}
