% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPpackage_split.R
\docType{data}
\name{IPpackage_split}
\alias{IPpackage_split}
\title{Split dos Dados Fictícios}
\format{
`IPpackage_split`
Um data frame com 6 linhas e 5 colunas:
\describe{
  \item{split_id}{Código de identificação do Split.}
  \item{split_nome}{Label de identificação do Split.}
  \item{split_tipo}{Identificação do tipo/nível do Split.}
  \item{filtro_quest}{Código SQL para filtrar o banco de dados.}
  \item{filtro_r}{Código R para filtrar o banco de dados.}
}
}
\source{
Fictício
}
\usage{
IPpackage_split
}
\description{
Split dos dados para simular um processamento.
}
\keyword{datasets}
