% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPpackage_split_peso.R
\docType{data}
\name{IPpackage_split_peso}
\alias{IPpackage_split_peso}
\title{split_peso dos Dados Fictícios}
\format{
`IPpackage_split_peso`
Um data frame com 6 linhas e 5 colunas:
\describe{
  \item{split_peso_id}{Código de identificação do split_peso.}
  \item{split_id}{Código de identificação do Split.}
  \item{questionario_id}{Código de identificação para join com o banco de dados (id do respondente).}
  \item{peso}{Peso.}
}
}
\source{
Fictício
}
\usage{
IPpackage_split_peso
}
\description{
split_peso dos dados para simular um processamento.
}
\keyword{datasets}
