% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PosProce_meanofmean.R
\name{PosProce_meanofmean}
\alias{PosProce_meanofmean}
\title{PosProce_meanofmean}
\usage{
PosProce_meanofmean(arquivo_medias, variaveis_recalcular, tipo = "Medias")
}
\arguments{
\item{arquivo_medias}{Aba 'Medias' ou 'Cruz_M' do processamento.}

\item{variaveis_recalcular}{Lista das variáveis a serem recalculadas,
acompanhadas de suas respectivas variáveis de composição (que serão incluídas
 no cálculo da média).}

\item{tipo}{Indicar a origem do 'arquivo_medias' (tipo = "Medias" ou tipo = "Cruz_M").}
}
\description{
A função tem a finalidade de recalcular a coluna 'media' e
'media_peso' do processamento. A abordagem utilizada implica em calcular a
média das variáveis especificadas no parâmetro 'variaveis_recalcular'
(uma lista) e substituir esses valores nos dados contidos no arquivo
referenciado pelo parâmetro 'arquivo_medias'. Este arquivo deve ser originado
do processamento.
}
\details{
Consute o livro para mais detalhes e exemplos.
}
\examples{

library(dplyr)
library(IPpackage)

# Data.frame com resultado das médias em um processamento
Medias <- base::data.frame(
 splits = base::rep(1:2, each = 6)
 ,variavel = base::rep(c("v29imrg","v29i", "v30i", "v31imrg", "v31i", "v32i"),
  times = 2)
 ,media = c(0, 71.02880658, 71.79012346, 0, 54.49834163, 53.64318567, 0,
  70.20933977, 73.51046699, 0, 57.48792271, 52.91311014)
 ,media_peso = c(0, 71.02880658, 71.79012346, 0, 54.49834163, 53.64318567,
  0, 70.20933977, 73.51046699, 0, 57.48792271, 52.91311014)
 ,n_base = rep(c(100, 38), each = 6)
 ,n_base_peso = rep(c(100, 38), each = 6)
 ,desvp = c(NA, 23.70668439, 23.40351373, NA, 29.08172754, 28.08884382, NA,
  21.18357424, 19.23669407, NA, 27.0142608, 24.55376985)
 ,desvp_peso = c(NA, 23.70668439, 23.40351373, NA, 29.08172754, 28.08884382,
  NA, 21.18357424, 19.23669407, NA, 27.0142608, 24.55376985)
 ,erro = c(NA, 1.018289287, 1.005266992, NA, 1.249167159, 1.206519152, NA,
  1.803265592, 1.637536146, NA, 2.299606593, 2.090155694)
 ,erro_peso = c(NA, 1.018289287, 1.005266992, NA, 1.249167159, 1.206519152,
  NA, 1.803265592, 1.637536146, NA, 2.299606593, 2.090155694)
 ,min = rep(0, times = 12)
 ,Q1 = c(NA, 55.55555556, 55.55555556, NA, 33.33333333, 44.44444444, NA,
  55.55555556, 66.66666667, NA, 44.44444444, 44.44444444)
 ,mediana = c(NA, 77.77777778, 77.77777778, NA, 55.55555556, 53.64318567,
  NA, 77.77777778, 77.77777778, NA, 66.66666667, 53.64318567)
 ,Q3 = c(NA, 88.88888889, 88.88888889, NA, 77.77777778, 77.77777778, NA,
  77.77777778, 88.88888889, NA, 77.77777778, 66.66666667)
 ,max = c(100, 100, 100, 100, 100, 100, 100, 100, 100, 100, 100, 100)
 ,cv = c(NA, 33.37615474, 32.59990734, NA, 53.36259173, 52.36237087, NA,
  30.17201744, 26.16864626, NA, 46.99119315, 46.40394373)
)

# Calculando média das médias com a função
novo_medias = IPpackage::PosProce_meanofmean(
 arquivo_medias = Medias,
 variaveis_recalcular = base::list(
   "v29imrg" = c("v29i", "v30i")
   ,"v31imrg" = c("v31i", "v32i")
 ),
 tipo = "Medias"
)

}
