% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_cor.R
\name{print_cor}
\alias{print_cor}
\title{print_cor}
\usage{
print_cor(
  df,
  cor,
  negrito = FALSE,
  italico = FALSE,
  borrado = FALSE,
  sublinhado = FALSE,
  data.frame = TRUE
)
}
\arguments{
\item{df}{Data.frame ou vetor}

\item{cor}{cor ='preto', 'vermelho', 'verde', 'amarelo', 'azul', 'roxo', 'ciano'
ou 'branco'}

\item{negrito}{negrito = 'TRUE' ou 'FALSE'}

\item{italico}{'italico = TRUE' ou 'FALSE'}

\item{borrado}{borrado = 'TRUE' ou 'FALSE'}

\item{sublinhado}{sublinhado = 'TRUE' ou 'FALSE'}

\item{data.frame}{data.frame = Imprimir no formato de data.frame. 'TRUE' ou 'FALSE'}
}
\description{
Imprimir um data.frame com uma cor específica
}
\details{
Consute o livro para mais detalhes e exemplos.
}
\examples{

IPpackage::print_cor(df="--------------show--------------",cor="verde",negrito=TRUE,
          italico=FALSE,borrado=FALSE,sublinhado=FALSE,data.frame=FALSE)
IPpackage::print_cor(df=c("a","b"),cor="ciano",negrito=TRUE,
          italico=TRUE,borrado=TRUE,sublinhado=TRUE,data.frame=FALSE)
IPpackage::print_cor(df=head(IPpackage::IPpackage_exemplo),cor="vermelho",data.frame=TRUE)

}
