% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_excel.R
\name{round_excel}
\alias{round_excel}
\title{round_excel}
\usage{
round_excel(y, digits = 0)
}
\arguments{
\item{y}{número ou vetor contendo números}

\item{digits}{número de casas decimais}
}
\description{
Arredonda um número (ou vetor contendo números) com saída semelhante ao Excel.
}
\details{
Consute o livro para mais detalhes e exemplos.
}
