% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_excel_decimais.R
\name{round_excel_decimais}
\alias{round_excel_decimais}
\title{round_excel_decimais}
\usage{
round_excel_decimais(y, n_depois_ponto = 0, ponto = ",")
}
\arguments{
\item{y}{número ou vetor contendo números}

\item{n_depois_ponto}{número de casas decimais}

\item{ponto}{separador decimal a ser retornado}
}
\description{
Muitas vezes, precisamos representar os resultados de um Excel em
 diferentes formas: tabelas em PNG, gráficos,... mas esbarramos com um pequeno
 problema: R e Excel utilizam métodos diferentes de arredondamento. Então, o
 que no excel vida 0.5, por exemplo, as saídas do R retornam 0.6.

 Para resolver esse problema, desenvolvemos a função round_excel, que arredonda
 os números seguindo a mesma metodologia que o Excel.
}
\details{
Consute o livro para mais detalhes e exemplos.
}
