% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substitui_prefixo_variavel.R
\name{substitui_prefixo_variavel}
\alias{substitui_prefixo_variavel}
\title{substitui_prefixo_variavel}
\usage{
substitui_prefixo_variavel(
  variavel,
  trocar_esse = "v",
  por_esse = "P",
  n_digitos = 3,
  add_p_n_digitos = "0",
  maiusculo = TRUE
)
}
\arguments{
\item{variavel}{Vetor contendo o(s) nome(s) das variáveis que devem ser
modificadas}

\item{trocar_esse}{String que especifica o prefixo que será substituído no(s)
nome(s) da(s) variável(is). O padrão é 'v'}

\item{por_esse}{String que será usada para substituir o valor especificado em
'trocar_esse'. O padrão é 'P'}

\item{n_digitos}{Número inteiro que define o total de dígitos desejado para
os nomes das variáveis após a substituição. O padrão é 3}

\item{add_p_n_digitos}{Caractere utilizado para preencher até que
o nome da variável tenha o número de dígitos especificado em 'n_digitos'.
O padrão é '0'}

\item{maiusculo}{Booleano que indica se os nomes das variáveis devem ser
convertidos para letras maiúsculas após a substituição. O padrão é TRUE.}
}
\description{
Esta função recebe um vetor contendo o(s) nome(s) de variável(is) que
começam com um valor especificado ('trocar_esse'). Ela substitui esse prefixo
inicial por um novo valor ('por_esse') e adiciona um caractere de preenchimento
('add_p_n_digitos') até que o nome tenha o número total de dígitos especificado
('n_digitos').

O uso mais comum desta função é substituir o prefixo 'v' ('trocar_esse' = 'v')
por 'P' ('por_esse' = 'P'), preenchendo com zeros ('add_p_n_digitos' = '0')
para que os nomes tenham exatamente 3 dígitos ('n_digitos' = 3).

Além disso, os nomes podem ser convertidos para letras maiúsculas
('maiusculo' = TRUE).
}
\details{
Consute o livro para mais detalhes e exemplos.
}
\examples{

# Utilizando o padrão
IPpackage::substitui_prefixo_variavel(
  c("v1", "v152", "v1521", "v5_3", "v5r1","v5mrg", "v2_1_asdfqwe_32")
)

# Modificando todos os parâmetros
IPpackage::substitui_prefixo_variavel(
  variavel = c("i1", "i152", "i1521", "i5_3", "i5r1","i5mrg", "i2_1_asdfqwe_32"),
  trocar_esse = "i",
  por_esse = "G",
  n_digitos = 5,
  add_p_n_digitos = "X",
  maiusculo = FALSE
)

}
