% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consistencia.R
\name{consistencia}
\alias{consistencia}
\title{consistencia}
\usage{
consistencia(
  log_consistenia = NULL,
  x,
  vars,
  nome = NULL,
  regra,
  pode_falta = FALSE,
  show = FALSE
)
}
\arguments{
\item{log_consistenia}{Se a variável em questão for a primeira a ser testada,
'log_consistenia' é definida como 'NULL'; no entanto,  se um arquivo no formato de
saída dessa função já existir,  'log_consistenia' será o seu data.frame correspondente}

\item{x}{Banco de dados}

\item{vars}{Variável ou,  no caso de um MRG,  vetor de variáveis}

\item{nome}{Se vars for um vetor -MRG,  este será o nome a ser atribuído a esse
conjunto de variáveis}

\item{regra}{Regra de entrada. Obs.: se a base for 100 porcento,  atribua o valor '100'}

\item{pode_falta}{'TRUE' para indicar a possibilidade de respostas ausentes;
'FALSE' para indicar a ausência de tolerância para respostas não fornecidas.}

\item{show}{Indicar as variáveis a serem impressas em caso de erro -serão
exibidas apenas as variáveis indicadas e as linhas com erros; 'FALSE' para não
imprimir as linhas,  mesmo na presença de erros.}
}
\description{
A função foi desenvolvida para avaliar se todos os indivíduos
designados para responder à variável/MRG cumpriram com essa atribuição.
}
\details{
Consute o livro para mais detalhes e exemplos.
}
\examples{

data=IPpackage::IPpackage_exemplo

##Tudo OK
log_consistenia = IPpackage::consistencia(
 log_consistenia = NULL,
 x = data,
 vars = "v1",
 regra = "100",
 pode_falta = FALSE,
 show = c("id", "v1")
)

log_consistenia = IPpackage::consistencia(
 log_consistenia = log_consistenia,
 x = data,
 vars = "v2",
 regra = "v1\%nin\%c(1)",
 pode_falta = FALSE,
 show = c("id", "v1", "v2")
)

log_consistenia = IPpackage::consistencia(
 log_consistenia = log_consistenia,
 x = data,
 vars = paste0("v3_", 1:3),
 regra = "v1\%nin\%c(1)",
 pode_falta = TRUE,
 show = c("id", "v1", "v2"),
 nome = "G003"
 )

log_consistenia = IPpackage::consistencia(
 log_consistenia = log_consistenia,
 x = data,
 vars = c("v4",  "v5",  "v6",  "v7",  "v8",  "v9",  "v10_1",  "v10_2",  "v11"),
 regra = "v1\%nin\%c(1)",
 pode_falta = TRUE,
 show = c("id", "v1", "v2")
)

log_consistenia = IPpackage::consistencia(
 log_consistenia = NULL,
 x = data,
 vars = "v1",
 regra = "100",
 pode_falta = FALSE,
 show = c("id", "v1", "v2")
)

log_consistenia = IPpackage::consistencia(
 log_consistenia = log_consistenia,
 x = data,
 vars = "v2",
 regra = "v1\%nin\%c(1)",
 pode_falta = FALSE,
 show = c("id", "v1", "v2")
)

log_consistenia = IPpackage::consistencia(
 log_consistenia = log_consistenia,
 x = data, vars = paste0("v3_", 1:3),
 regra = "v1\%nin\%c(1)",
 pode_falta = TRUE,
 show = c("id", "v1", paste0("v3_", 1:3)),
 nome = "G003"
 )

}
