% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUN_cruzaMRG.R
\name{FUN_cruzaMRG}
\alias{FUN_cruzaMRG}
\title{FUN_cruzaMRG}
\usage{
FUN_cruzaMRG(TABELA, DICIONARIO, var1, varMRG, var_fecha)
}
\arguments{
\item{TABELA}{Banco de dados a ser analisado.}

\item{DICIONARIO}{Dicionário correspondente à TABELA.}

\item{var1}{Primeira variável do cruzamento. Pode ser MRG ou variável isolada (lista).}

\item{varMRG}{Segunda variável do cruzamento. Tem que ser um MRG (lista).}

\item{var_fecha}{Variável na coluna.}
}
\description{
Calcula duas possibilidades de frequências cruzadas: 'var vs mrg' ou 'mrg vs mrg'.
O parâmetro 'var_fecha' define a variável na coluna.
}
\details{
Consute o livro para mais detalhes e exemplos.
}
\examples{

library(dplyr)
library(tidyr)
library(stringr)
library(IPpackage)

# Var isolada vs mrg fechando na var isolada----------------------------------
IPpackage::FUN_cruzaMRG(
 TABELA = IPpackage::IPpackage_exemplo,
 DICIONARIO = IPpackage::IPpackage_dicionario,
 var1 = base::list("v1"="v1"),
 varMRG = base::list("v3mrg" = c("v3_1","v3_2","v3_3")),
 var_fecha = "v1"
)

# Var isolada vs mrg fechando no mrg------------------------------------------
IPpackage::FUN_cruzaMRG(
 TABELA = IPpackage::IPpackage_exemplo,
 DICIONARIO = IPpackage::IPpackage_dicionario,
 var1 = base::list("v1"="v1"),
 varMRG = base::list("v3mrg" = c("v3_1","v3_2","v3_3")),
 var_fecha = "v3mrg"
)
# mrg vs mrg------------------------------------------------------------------

# Fechando em 'v3mrg
IPpackage::FUN_cruzaMRG(
 TABELA = IPpackage::IPpackage_exemplo,
 DICIONARIO = IPpackage::IPpackage_dicionario,
 var1 = base::list("v19mrg"=c("v19_1","v19_2")),
 varMRG = base::list("v3mrg" = c("v3_1","v3_2","v3_3")),
 var_fecha = "v3mrg"
)

# Fechando em 'v19mrg
IPpackage::FUN_cruzaMRG(
 TABELA = IPpackage::IPpackage_exemplo,
 DICIONARIO = IPpackage::IPpackage_dicionario,
 var1 = base::list("v19mrg"=c("v19_1","v19_2")),
 varMRG = base::list("v3mrg" = c("v3_1","v3_2","v3_3")),
 var_fecha = "v19mrg"
)

}
