% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPpackage_exemplo.R
\docType{data}
\name{IPpackage_exemplo}
\alias{IPpackage_exemplo}
\title{Dados Fictícios}
\format{
`IPpackage_exemplo`
Um data frame com 100 linhas e X colunas:
\describe{
  \item{id}{ID}
  \item{v1}{Opções}
  \item{v2}{Numérico (só responde quem não respondeu v1==1)}
  \item{v3_1,v3_2,v3_3}{MRG (só responde quem não respondeu v1==1)}
  \item{v4,v5,v6,v7,v8,v9,v10,v11}{MRG citou; v10=Outros; v11=NS/NR
  (só responde quem não respondeu v1==1)}
  \item{v12, v13}{Variáveis de escala}
  ...
}
}
\source{
Fictício
}
\usage{
IPpackage_exemplo
}
\description{
Dados que simulam um processamento
}
\keyword{datasets}
